ref: rename Rooms to RoomsView (only filename, class already)
